add jdbc

+++ src/main/java/productpackage/BankProduct.java
@@ -2,15 +2,16 @@ package productpackage;
 
 import ATMpackage.Transaction;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface BankProduct {
     String getNumber();
     String getCurrency();
-    double getBalance();
+    BigDecimal getBalance();
     List<Transaction> getTransactionList();
 
-    void setBalance(double d);
+    void setBalance(BigDecimal d);
     void setTransactionList(List<Transaction> list);
 
 }
