+++ src/main/java/productpackage/Deposit.java
@@ -5,6 +5,7 @@ import ATMpackage.Transaction;
 import lombok.*;
 
 import javax.validation.constraints.Min;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,10 +16,10 @@ public class Deposit implements BankProduct {
         private final String number;
         private final String currency;
         @Min(0)
-        private double balance;             //:todo изменить на бигдесимал
+        private BigDecimal balance;             //:todo изменить на бигдесимал
         private List<Transaction> transactionList;
 
-        public Deposit(String number, String currency, double balance) {
+        public Deposit(String number, String currency, BigDecimal balance) {
                 this.number = number;
                 this.currency = currency;
                 this.balance = balance;
